Environments with validator types
