+++ rust/Cargo.toml
@@ -15,6 +15,6 @@ alloy-sol-types.workspace = true
 rand_xoshiro.workspace = true
 rand.workspace = true
 
-pyo3 = { version="0.20.0", features = ["extension-module"] }
+pyo3 = { version="0.20.3", features = ["extension-module"] }
 
 verbs-rs = { path = "../crates/verbs_rs" }
