+++ rust/src/sim/mod.rs
@@ -1,7 +1,8 @@
 mod base_env;
 mod empty_env;
 mod fork_env;
+mod interface_macro;
 mod snapshot;
 
-pub use empty_env::EmptyEnv;
-pub use fork_env::ForkEnv;
+pub use empty_env::{EmptyEnvGasPriority, EmptyEnvRandom};
+pub use fork_env::{ForkEnvGasPriority, ForkEnvRandom};
